removed tutorial level and associated hints
The only things that need explanation as the game stands are the
controls and the character classes. This is done in the character
selection level.

The rest of the game *must* be self explanatory when the player first
sees it or experiences it. If not, there is something wrong in the
design/gfx/sfx.